refactor: change module name for smoke tests
