--- sofa-boot-tests/sofa-boot-smoke-tests/sofa-boot-smoke-tests-integration-test/src/test/java/com/alipay/sofa/smoke/tests/integration/test/stub/ExternalServiceClientImpl.java
@@ -1,40 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.alipay.sofa.smoke.tests.integration.test.stub;
-
-import com.alipay.sofa.runtime.api.annotation.SofaService;
-import com.alipay.sofa.smoke.tests.integration.test.ExternalServiceClient;
-import jakarta.annotation.PostConstruct;
-import org.springframework.stereotype.Service;
-
-/**
- * @author pengym
- * @version ExternalServiceClientA.java, v 0.1 2023年08月08日 15:53 pengym
- */
-@SofaService(interfaceType = ExternalServiceClient.class)
-@Service
-public class ExternalServiceClientImpl implements ExternalServiceClient {
-    @PostConstruct
-    public void postConstruct() {
-        System.out.println("ExternalServiceClientA has been constructed");
-    }
-
-    @Override
-    public String invoke(Object... args) {
-        throw new UnsupportedOperationException();
-    }
-}
